perf(bundle): reduce main bundle size with code splitting
- Lazy-load date-fns locales (de, fr, it) instead of bundling all upfront
- Split i18n translations into separate files for lazy loading
- Add manual chunks for vendor code (react, router, state, zod)
- Update size-limit config to track main bundle and vendor chunks separately

Bundle size improvements:
- Main bundle: 162 KB → 107 KB gzipped (34% reduction)
- Locales now lazy-loaded: ~18 KB moved to on-demand chunks
- Vendor chunks separated for better caching

Initial load for English users now ~153 KB vs previous ~162 KB.
Non-English users load their locale chunk (~5 KB) on-demand.
