+++ web-app/src/i18n/locales/de.ts
@@ -1,0 +1,259 @@
+import type { Translations } from "../types";
+
+const de: Translations = {
+  common: {
+    loading: "Laden...",
+    error: "Ein Fehler ist aufgetreten",
+    retry: "Erneut versuchen",
+    cancel: "Abbrechen",
+    save: "Speichern",
+    close: "Schliessen",
+    confirm: "Bestätigen",
+    noResults: "Keine Ergebnisse gefunden",
+    today: "Heute",
+    tomorrow: "Morgen",
+    home: "Heim",
+    away: "Gast",
+    men: "Herren",
+    women: "Damen",
+    match: "Spiel",
+    dateTime: "Datum & Zeit",
+    location: "Ort",
+    position: "Position",
+    requiredLevel: "Erforderliches Niveau",
+    demoModeBanner: "Demo-Modus - Beispieldaten werden angezeigt",
+    optional: "Optional",
+  },
+  auth: {
+    login: "Anmelden",
+    logout: "Abmelden",
+    username: "Benutzername",
+    password: "Passwort",
+    loginButton: "Anmelden",
+    loggingIn: "Anmeldung...",
+    invalidCredentials: "Ungültiger Benutzername oder Passwort",
+    sessionExpired: "Sitzung abgelaufen. Bitte erneut anmelden.",
+    subtitle: "Schweizer Volleyball Schiedsrichter Management",
+    or: "oder",
+    demoMode: "Demo-Modus ausprobieren",
+    loginInfo: "Verwenden Sie Ihre VolleyManager-Anmeldeinformationen.",
+    privacyNote: "Ihr Passwort wird niemals gespeichert.",
+    loadingDemo: "Demo-Modus wird geladen...",
+  },
+  occupations: {
+    referee: "Schiedsrichter",
+    player: "Spieler",
+    clubAdmin: "Vereinsadministrator",
+    associationAdmin: "Verband",
+  },
+  assignments: {
+    title: "Einsätze",
+    upcoming: "Bevorstehend",
+    past: "Vergangen",
+    validationClosed: "Validierung geschlossen",
+    loading: "Einsätze werden geladen...",
+    noAssignments: "Keine Einsätze gefunden",
+    noUpcomingTitle: "Keine bevorstehenden Einsätze",
+    noUpcomingDescription:
+      "Sie haben keine bevorstehenden Schiedsrichtereinsätze geplant.",
+    noClosedTitle: "Keine abgeschlossenen Einsätze",
+    noClosedDescription:
+      "Keine Einsätze mit abgeschlossener Validierung in dieser Saison.",
+    confirmed: "Bestätigt",
+    pending: "Ausstehend",
+    cancelled: "Abgesagt",
+    editCompensation: "Entschädigung bearbeiten",
+    validateGame: "Spieldetails validieren",
+    kilometers: "Kilometer",
+    reason: "Grund",
+    reasonPlaceholder: "Grund für Änderung der Entschädigung eingeben",
+    homeScore: "Heimscore",
+    awayScore: "Gastscore",
+    numberOfSets: "Anzahl Sätze",
+    gameReportNotAvailable:
+      "Spielberichte sind nur für NLA- und NLB-Spiele verfügbar.",
+  },
+  compensations: {
+    title: "Entschädigungen",
+    noCompensations: "Keine Entschädigungen gefunden",
+    paid: "Bezahlt",
+    unpaid: "Unbezahlt",
+    pending: "Ausstehend",
+    total: "Total",
+    all: "Alle",
+    received: "Erhalten",
+    loading: "Entschädigungen werden geladen...",
+    noCompensationsTitle: "Keine Entschädigungen",
+    noCompensationsDescription: "Sie haben noch keine Entschädigungseinträge.",
+    noPaidTitle: "Keine bezahlten Entschädigungen",
+    noPaidDescription: "Keine bezahlten Entschädigungen gefunden.",
+    noUnpaidTitle: "Keine ausstehenden Entschädigungen",
+    noUnpaidDescription: "Keine ausstehenden Entschädigungen. Alles erledigt!",
+    errorLoading: "Entschädigungen konnten nicht geladen werden",
+    pdfNotAvailableDemo: "PDF-Downloads sind im Demo-Modus nicht verfügbar",
+    pdfDownloadFailed:
+      "PDF konnte nicht heruntergeladen werden. Bitte versuchen Sie es später erneut.",
+  },
+  exchange: {
+    title: "Tauschbörse",
+    noExchanges: "Keine Tauschangebote verfügbar",
+    apply: "Bewerben",
+    withdraw: "Zurückziehen",
+    open: "Offen",
+    applied: "Beworben",
+    closed: "Geschlossen",
+    all: "Alle",
+    myApplications: "Meine Bewerbungen",
+    takeOverTitle: "Einsatz übernehmen",
+    takeOverConfirm: "Möchten Sie diesen Einsatz wirklich übernehmen?",
+    takeOverButton: "Übernahme bestätigen",
+    removeTitle: "Aus Tauschbörse entfernen",
+    removeConfirm:
+      "Möchten Sie diesen Einsatz wirklich aus der Tauschbörse entfernen?",
+    removeButton: "Aus Tauschbörse entfernen",
+    filterByLevel: "Nur mein Niveau",
+    noExchangesAtLevel: "Keine Tauschangebote auf Ihrem Niveau verfügbar.",
+    noOpenExchangesTitle: "Keine offenen Tauschangebote",
+    noOpenExchangesDescription:
+      "Derzeit sind keine Schiedsrichterpositionen zum Tausch verfügbar.",
+    noApplicationsTitle: "Keine Bewerbungen",
+    noApplicationsDescription:
+      "Sie haben sich noch für keine Tauschangebote beworben.",
+  },
+  positions: {
+    "head-one": "1. Schiedsrichter",
+    "head-two": "2. Schiedsrichter",
+    "linesman-one": "Linienrichter 1",
+    "linesman-two": "Linienrichter 2",
+    "linesman-three": "Linienrichter 3",
+    "linesman-four": "Linienrichter 4",
+    "standby-head": "Ersatz Schiedsrichter",
+    "standby-linesman": "Ersatz Linienrichter",
+  },
+  nav: {
+    assignments: "Einsätze",
+    compensations: "Entschädigungen",
+    exchange: "Tauschbörse",
+    settings: "Einstellungen",
+  },
+  settings: {
+    title: "Einstellungen",
+    profile: "Profil",
+    language: "Sprache",
+    safeMode: "Sicherheitsmodus",
+    safeModeDescription:
+      "Der Sicherheitsmodus beschränkt gefährliche Operationen wie das Hinzufügen/Übernehmen von Spielen zur/von Tauschbörse oder das Validieren von Spielen. Dies hilft, versehentliche Änderungen während des Testens der App zu vermeiden.",
+    safeModeEnabled: "Sicherheitsmodus ist aktiviert",
+    safeModeDisabled: "Sicherheitsmodus ist deaktiviert",
+    safeModeWarningTitle: "Sicherheitsmodus deaktivieren?",
+    safeModeWarningMessage:
+      "Das Deaktivieren des Sicherheitsmodus ermöglicht Operationen, die Ihre Einsätze und Spiele ändern können.",
+    safeModeWarningPoint1:
+      "volleymanager.volleyball.ch ist die einzige massgebende Quelle",
+    safeModeWarningPoint2:
+      "Überprüfen Sie Ihre Änderungen immer auf der offiziellen VolleyManager-Website",
+    safeModeWarningPoint3:
+      "VolleyKit übernimmt keine Verantwortung für allfällige Fehler",
+    safeModeConfirmButton: "Ich verstehe, deaktivieren",
+    safeModeDangerous: "Gefährliche Operationen sind aktiviert",
+    safeModeBlocked:
+      "Diese Operation ist im Sicherheitsmodus gesperrt. Deaktivieren Sie den Sicherheitsmodus in den Einstellungen, um fortzufahren.",
+    privacy: "Datenschutz",
+    privacyNoCollection:
+      "VolleyKit sammelt oder speichert keine persönlichen Daten.",
+    privacyDirectComm:
+      "Alle Daten fliessen direkt zwischen Ihrem Browser und den Servern von Swiss Volley.",
+    privacyNoAnalytics: "Kein Tracking, keine Analysen, keine Telemetrie.",
+    about: "Über",
+    version: "Version",
+    platform: "Plattform",
+    openWebsite: "VolleyManager-Website öffnen",
+    roles: "Rollen",
+    dataSource: "Daten von volleymanager.volleyball.ch",
+    disclaimer:
+      "Inoffizielle App für den persönlichen Gebrauch. Alle Daten sind Eigentum von Swiss Volley.",
+    updates: "Updates",
+    checkForUpdates: "Nach Updates suchen",
+    checking: "Überprüfen...",
+    upToDate: "App ist aktuell",
+    updateAvailable: "Update verfügbar",
+    lastChecked: "Zuletzt geprüft",
+    updateNow: "Jetzt aktualisieren",
+    updateCheckFailed: "Update-Prüfung fehlgeschlagen",
+  },
+  validation: {
+    homeRoster: "Heimkader",
+    awayRoster: "Gastkader",
+    scorer: "Schreiber",
+    scoresheet: "Spielbericht",
+    homeRosterPlaceholder:
+      "Die Überprüfung des Heimkaders wird hier verfügbar sein.",
+    awayRosterPlaceholder:
+      "Die Überprüfung des Gastkaders wird hier verfügbar sein.",
+    scorerPlaceholder: "Die Schreiberidentifikation wird hier verfügbar sein.",
+    scoresheetPlaceholder:
+      "Der Upload des Spielberichts wird hier verfügbar sein.",
+    addPlayer: "Spieler hinzufügen",
+    searchPlayers: "Spieler suchen...",
+    noPlayersFound: "Keine Spieler gefunden",
+    loadPlayersError: "Spieler konnten nicht geladen werden",
+    playerAlreadyAdded: "Bereits im Kader",
+    jerseyNumber: "Trikot #",
+    license: "Lizenz",
+    roster: {
+      addPlayer: "Spieler hinzufügen",
+      removePlayer: "Spieler entfernen",
+      undoRemoval: "Entfernung rückgängig",
+      newlyAdded: "Neu",
+      captain: "Kapitän",
+      libero: "Libero",
+      emptyRoster: "Keine Spieler im Kader",
+      loadingRoster: "Kader wird geladen...",
+      errorLoading: "Kader konnte nicht geladen werden",
+      playerCount: "{count} Spieler",
+    },
+    scorerSearch: {
+      searchPlaceholder: "Schreiber nach Name suchen...",
+      searchHint:
+        "Namen eingeben (z.B. 'Müller' oder 'Hans Müller') oder Geburtsjahr hinzufügen (z.B. 'Müller 1985')",
+      searchError: "Schreibersuche fehlgeschlagen",
+      noScorerSelected:
+        "Kein Schreiber ausgewählt. Verwenden Sie die Suche oben, um einen Schreiber zu finden und auszuwählen.",
+      noScorersFound: "Keine Schreiber gefunden",
+      searchResults: "Suchergebnisse",
+      resultsCount: "{count} Ergebnisse gefunden",
+      resultsCountOne: "1 Ergebnis gefunden",
+    },
+    scoresheetUpload: {
+      title: "Spielbericht hochladen",
+      description:
+        "Laden Sie ein Foto oder einen Scan des physischen Spielberichts hoch",
+      acceptedFormats: "JPEG, PNG oder PDF",
+      maxFileSize: "Max. 10 MB",
+      selectFile: "Datei auswählen",
+      takePhoto: "Foto aufnehmen",
+      uploading: "Wird hochgeladen...",
+      uploadComplete: "Upload abgeschlossen",
+      replace: "Ersetzen",
+      remove: "Entfernen",
+      fileTooLarge: "Datei ist zu gross. Maximale Grösse ist 10 MB.",
+      invalidFileType:
+        "Ungültiger Dateityp. Bitte verwenden Sie JPEG, PNG oder PDF.",
+      demoModeNote: "Demo-Modus: Uploads werden simuliert",
+      previewAlt: "Spielbericht-Vorschau",
+    },
+    state: {
+      unsavedChangesTitle: "Ungespeicherte Änderungen",
+      unsavedChangesMessage:
+        "Sie haben ungespeicherte Änderungen. Möchten Sie diese wirklich verwerfen?",
+      continueEditing: "Weiter bearbeiten",
+      discardChanges: "Änderungen verwerfen",
+      saveSuccess: "Validierung erfolgreich gespeichert",
+      saveError: "Validierung konnte nicht gespeichert werden",
+      saveDisabledTooltip:
+        "Füllen Sie alle erforderlichen Bereiche aus (Heimkader, Gastkader, Schreiber), um zu speichern",
+    },
+  },
+};
+
+export default de;
