+++ web-app/src/i18n/locales/en.ts
@@ -1,0 +1,255 @@
+import type { Translations } from "../types";
+
+const en: Translations = {
+  common: {
+    loading: "Loading...",
+    error: "An error occurred",
+    retry: "Retry",
+    cancel: "Cancel",
+    save: "Save",
+    close: "Close",
+    confirm: "Confirm",
+    noResults: "No results found",
+    today: "Today",
+    tomorrow: "Tomorrow",
+    home: "Home",
+    away: "Away",
+    men: "Men",
+    women: "Women",
+    match: "Match",
+    dateTime: "Date & Time",
+    location: "Location",
+    position: "Position",
+    requiredLevel: "Required Level",
+    demoModeBanner: "Demo Mode - Viewing sample data",
+    optional: "Optional",
+  },
+  auth: {
+    login: "Login",
+    logout: "Logout",
+    username: "Username",
+    password: "Password",
+    loginButton: "Sign in",
+    loggingIn: "Signing in...",
+    invalidCredentials: "Invalid username or password",
+    sessionExpired: "Session expired. Please log in again.",
+    subtitle: "Swiss Volleyball Referee Management",
+    or: "or",
+    demoMode: "Try Demo Mode",
+    loginInfo: "Use your VolleyManager credentials to login.",
+    privacyNote: "Your password is never stored.",
+    loadingDemo: "Loading demo mode...",
+  },
+  occupations: {
+    referee: "Referee",
+    player: "Player",
+    clubAdmin: "Club Admin",
+    associationAdmin: "Association",
+  },
+  assignments: {
+    title: "Assignments",
+    upcoming: "Upcoming",
+    past: "Past",
+    validationClosed: "Validation Closed",
+    loading: "Loading assignments...",
+    noAssignments: "No assignments found",
+    noUpcomingTitle: "No upcoming assignments",
+    noUpcomingDescription:
+      "You have no upcoming referee assignments scheduled.",
+    noClosedTitle: "No closed assignments",
+    noClosedDescription:
+      "No assignments with closed validation in this season.",
+    confirmed: "Confirmed",
+    pending: "Pending",
+    cancelled: "Cancelled",
+    editCompensation: "Edit Compensation",
+    validateGame: "Validate Game Details",
+    kilometers: "Kilometers",
+    reason: "Reason",
+    reasonPlaceholder: "Enter reason for compensation change",
+    homeScore: "Home Score",
+    awayScore: "Away Score",
+    numberOfSets: "Number of Sets",
+    gameReportNotAvailable:
+      "Game reports are only available for NLA and NLB games.",
+  },
+  compensations: {
+    title: "Compensations",
+    noCompensations: "No compensations found",
+    paid: "Paid",
+    unpaid: "Unpaid",
+    pending: "Pending",
+    total: "Total",
+    all: "All",
+    received: "Received",
+    loading: "Loading compensations...",
+    noCompensationsTitle: "No compensations",
+    noCompensationsDescription: "You have no compensation records yet.",
+    noPaidTitle: "No paid compensations",
+    noPaidDescription: "No paid compensations found.",
+    noUnpaidTitle: "No pending compensations",
+    noUnpaidDescription: "No pending compensations. All caught up!",
+    errorLoading: "Failed to load compensations",
+    pdfNotAvailableDemo: "PDF downloads are not available in demo mode",
+    pdfDownloadFailed:
+      "Failed to download compensation PDF. Please try again later.",
+  },
+  exchange: {
+    title: "Exchange",
+    noExchanges: "No exchanges available",
+    apply: "Apply",
+    withdraw: "Withdraw",
+    open: "Open",
+    applied: "Applied",
+    closed: "Closed",
+    all: "All",
+    myApplications: "My Applications",
+    takeOverTitle: "Take Over Assignment",
+    takeOverConfirm: "Are you sure you want to take over this assignment?",
+    takeOverButton: "Confirm Take Over",
+    removeTitle: "Remove from Exchange",
+    removeConfirm:
+      "Are you sure you want to remove this assignment from the exchange?",
+    removeButton: "Remove from Exchange",
+    filterByLevel: "My level only",
+    noExchangesAtLevel: "No exchanges available at your level.",
+    noOpenExchangesTitle: "No open exchanges",
+    noOpenExchangesDescription:
+      "There are currently no referee positions available for exchange.",
+    noApplicationsTitle: "No applications",
+    noApplicationsDescription: "You haven't applied for any exchanges yet.",
+  },
+  positions: {
+    "head-one": "1st Referee",
+    "head-two": "2nd Referee",
+    "linesman-one": "Linesman 1",
+    "linesman-two": "Linesman 2",
+    "linesman-three": "Linesman 3",
+    "linesman-four": "Linesman 4",
+    "standby-head": "Standby Head",
+    "standby-linesman": "Standby Linesman",
+  },
+  nav: {
+    assignments: "Assignments",
+    compensations: "Compensations",
+    exchange: "Exchange",
+    settings: "Settings",
+  },
+  settings: {
+    title: "Settings",
+    profile: "Profile",
+    language: "Language",
+    safeMode: "Safe Mode",
+    safeModeDescription:
+      "Safe mode restricts dangerous operations like adding/taking games from exchange or validating games. This helps prevent accidental modifications while the app is being tested.",
+    safeModeEnabled: "Safe mode is enabled",
+    safeModeDisabled: "Safe mode is disabled",
+    safeModeWarningTitle: "Disable Safe Mode?",
+    safeModeWarningMessage:
+      "Disabling safe mode will enable operations that may modify your assignments and games.",
+    safeModeWarningPoint1:
+      "volleymanager.volleyball.ch is the only authoritative source of truth",
+    safeModeWarningPoint2:
+      "Always verify your changes on the official VolleyManager website",
+    safeModeWarningPoint3:
+      "VolleyKit takes no responsibility for any errors that may occur",
+    safeModeConfirmButton: "I Understand, Disable",
+    safeModeDangerous: "Dangerous operations are enabled",
+    safeModeBlocked:
+      "This operation is blocked in safe mode. Disable safe mode in Settings to proceed.",
+    privacy: "Privacy",
+    privacyNoCollection:
+      "VolleyKit does not collect or store any personal data.",
+    privacyDirectComm:
+      "All data flows directly between your browser and Swiss Volley's servers.",
+    privacyNoAnalytics: "No tracking, analytics, or telemetry.",
+    about: "About",
+    version: "Version",
+    platform: "Platform",
+    openWebsite: "Open VolleyManager website",
+    roles: "Roles",
+    dataSource: "Data from volleymanager.volleyball.ch",
+    disclaimer:
+      "Unofficial app for personal use. All data is property of Swiss Volley.",
+    updates: "Updates",
+    checkForUpdates: "Check for Updates",
+    checking: "Checking...",
+    upToDate: "App is up to date",
+    updateAvailable: "Update available",
+    lastChecked: "Last checked",
+    updateNow: "Update Now",
+    updateCheckFailed: "Update check failed",
+  },
+  validation: {
+    homeRoster: "Home Roster",
+    awayRoster: "Away Roster",
+    scorer: "Scorer",
+    scoresheet: "Scoresheet",
+    homeRosterPlaceholder:
+      "Home team roster verification will be available here.",
+    awayRosterPlaceholder:
+      "Away team roster verification will be available here.",
+    scorerPlaceholder: "Scorer identification will be available here.",
+    scoresheetPlaceholder: "Scoresheet upload will be available here.",
+    addPlayer: "Add Player",
+    searchPlayers: "Search players...",
+    noPlayersFound: "No players found",
+    loadPlayersError: "Failed to load players",
+    playerAlreadyAdded: "Already on roster",
+    jerseyNumber: "Jersey #",
+    license: "License",
+    roster: {
+      addPlayer: "Add Player",
+      removePlayer: "Remove player",
+      undoRemoval: "Undo removal",
+      newlyAdded: "New",
+      captain: "Captain",
+      libero: "Libero",
+      emptyRoster: "No players in roster",
+      loadingRoster: "Loading roster...",
+      errorLoading: "Failed to load roster",
+      playerCount: "{count} players",
+    },
+    scorerSearch: {
+      searchPlaceholder: "Search scorer by name...",
+      searchHint:
+        "Enter name (e.g., 'Müller' or 'Hans Müller') or add birth year (e.g., 'Müller 1985')",
+      searchError: "Failed to search scorers",
+      noScorerSelected:
+        "No scorer selected. Use the search above to find and select a scorer.",
+      noScorersFound: "No scorers found",
+      searchResults: "Search results",
+      resultsCount: "{count} results found",
+      resultsCountOne: "1 result found",
+    },
+    scoresheetUpload: {
+      title: "Upload Scoresheet",
+      description: "Upload a photo or scan of the physical scoresheet",
+      acceptedFormats: "JPEG, PNG, or PDF",
+      maxFileSize: "Max 10 MB",
+      selectFile: "Select File",
+      takePhoto: "Take Photo",
+      uploading: "Uploading...",
+      uploadComplete: "Upload complete",
+      replace: "Replace",
+      remove: "Remove",
+      fileTooLarge: "File is too large. Maximum size is 10 MB.",
+      invalidFileType: "Invalid file type. Please use JPEG, PNG, or PDF.",
+      demoModeNote: "Demo mode: uploads are simulated",
+      previewAlt: "Scoresheet preview",
+    },
+    state: {
+      unsavedChangesTitle: "Unsaved Changes",
+      unsavedChangesMessage:
+        "You have unsaved changes. Are you sure you want to discard them?",
+      continueEditing: "Continue Editing",
+      discardChanges: "Discard Changes",
+      saveSuccess: "Validation saved successfully",
+      saveError: "Failed to save validation",
+      saveDisabledTooltip:
+        "Complete all required panels (Home Roster, Away Roster, Scorer) to save",
+    },
+  },
+};
+
+export default en;
