+++ web-app/src/i18n/locales/it.ts
@@ -1,0 +1,254 @@
+import type { Translations } from "../types";
+
+const it: Translations = {
+  common: {
+    loading: "Caricamento...",
+    error: "Si è verificato un errore",
+    retry: "Riprova",
+    cancel: "Annulla",
+    save: "Salva",
+    close: "Chiudi",
+    confirm: "Conferma",
+    noResults: "Nessun risultato trovato",
+    today: "Oggi",
+    tomorrow: "Domani",
+    home: "Casa",
+    away: "Ospite",
+    men: "Uomini",
+    women: "Donne",
+    match: "Partita",
+    dateTime: "Data e Ora",
+    location: "Luogo",
+    position: "Posizione",
+    requiredLevel: "Livello richiesto",
+    demoModeBanner: "Modalità Demo - Dati di esempio",
+    optional: "Opzionale",
+  },
+  auth: {
+    login: "Accesso",
+    logout: "Esci",
+    username: "Nome utente",
+    password: "Password",
+    loginButton: "Accedi",
+    loggingIn: "Accesso in corso...",
+    invalidCredentials: "Nome utente o password non validi",
+    sessionExpired: "Sessione scaduta. Effettua nuovamente il login.",
+    subtitle: "Gestione arbitri di pallavolo svizzera",
+    or: "o",
+    demoMode: "Prova la modalità demo",
+    loginInfo: "Usa le tue credenziali VolleyManager per accedere.",
+    privacyNote: "La tua password non viene mai memorizzata.",
+    loadingDemo: "Caricamento modalità demo...",
+  },
+  occupations: {
+    referee: "Arbitro",
+    player: "Giocatore",
+    clubAdmin: "Admin club",
+    associationAdmin: "Associazione",
+  },
+  assignments: {
+    title: "Designazioni",
+    upcoming: "In programma",
+    past: "Passate",
+    validationClosed: "Validazione chiusa",
+    loading: "Caricamento designazioni...",
+    noAssignments: "Nessuna designazione trovata",
+    noUpcomingTitle: "Nessuna designazione in programma",
+    noUpcomingDescription: "Non hai designazioni arbitrali in programma.",
+    noClosedTitle: "Nessuna designazione chiusa",
+    noClosedDescription:
+      "Nessuna designazione con validazione chiusa in questa stagione.",
+    confirmed: "Confermato",
+    pending: "In attesa",
+    cancelled: "Annullato",
+    editCompensation: "Modifica compenso",
+    validateGame: "Valida dettagli partita",
+    kilometers: "Chilometri",
+    reason: "Motivo",
+    reasonPlaceholder: "Inserisci motivo per cambio compenso",
+    homeScore: "Punteggio casa",
+    awayScore: "Punteggio ospite",
+    numberOfSets: "Numero di set",
+    gameReportNotAvailable:
+      "I rapporti delle partite sono disponibili solo per le partite NLA e NLB.",
+  },
+  compensations: {
+    title: "Compensi",
+    noCompensations: "Nessun compenso trovato",
+    paid: "Pagato",
+    unpaid: "Non pagato",
+    pending: "In attesa",
+    total: "Totale",
+    all: "Tutti",
+    received: "Ricevuto",
+    loading: "Caricamento compensi...",
+    noCompensationsTitle: "Nessun compenso",
+    noCompensationsDescription: "Non hai ancora voci di compenso.",
+    noPaidTitle: "Nessun compenso pagato",
+    noPaidDescription: "Nessun compenso pagato trovato.",
+    noUnpaidTitle: "Nessun compenso in attesa",
+    noUnpaidDescription: "Nessun compenso in attesa. Tutto aggiornato!",
+    errorLoading: "Impossibile caricare i compensi",
+    pdfNotAvailableDemo: "I download PDF non sono disponibili in modalità demo",
+    pdfDownloadFailed: "Impossibile scaricare il PDF. Riprova più tardi.",
+  },
+  exchange: {
+    title: "Borsa scambi",
+    noExchanges: "Nessuno scambio disponibile",
+    apply: "Candidati",
+    withdraw: "Ritira",
+    open: "Aperto",
+    applied: "Candidato",
+    closed: "Chiuso",
+    all: "Tutti",
+    myApplications: "Le mie candidature",
+    takeOverTitle: "Assumere la designazione",
+    takeOverConfirm: "Sei sicuro di voler assumere questa designazione?",
+    takeOverButton: "Conferma assunzione",
+    removeTitle: "Rimuovere dalla borsa",
+    removeConfirm:
+      "Sei sicuro di voler rimuovere questa designazione dalla borsa?",
+    removeButton: "Rimuovere dalla borsa",
+    filterByLevel: "Solo il mio livello",
+    noExchangesAtLevel: "Nessuno scambio disponibile al tuo livello.",
+    noOpenExchangesTitle: "Nessuno scambio aperto",
+    noOpenExchangesDescription:
+      "Al momento non ci sono posizioni arbitrali disponibili per lo scambio.",
+    noApplicationsTitle: "Nessuna candidatura",
+    noApplicationsDescription:
+      "Non hai ancora fatto domanda per nessuno scambio.",
+  },
+  positions: {
+    "head-one": "1° Arbitro",
+    "head-two": "2° Arbitro",
+    "linesman-one": "Giudice di linea 1",
+    "linesman-two": "Giudice di linea 2",
+    "linesman-three": "Giudice di linea 3",
+    "linesman-four": "Giudice di linea 4",
+    "standby-head": "Arbitro riserva",
+    "standby-linesman": "Giudice di linea riserva",
+  },
+  nav: {
+    assignments: "Designazioni",
+    compensations: "Compensi",
+    exchange: "Scambi",
+    settings: "Impostazioni",
+  },
+  settings: {
+    title: "Impostazioni",
+    profile: "Profilo",
+    language: "Lingua",
+    safeMode: "Modalità sicura",
+    safeModeDescription:
+      "La modalità sicura limita operazioni pericolose come l'aggiunta/assunzione di partite dalla borsa scambi o la convalida di partite. Questo aiuta a prevenire modifiche accidentali durante il test dell'app.",
+    safeModeEnabled: "La modalità sicura è attivata",
+    safeModeDisabled: "La modalità sicura è disattivata",
+    safeModeWarningTitle: "Disattivare la modalità sicura?",
+    safeModeWarningMessage:
+      "La disattivazione della modalità sicura abiliterà operazioni che potrebbero modificare le tue designazioni e partite.",
+    safeModeWarningPoint1:
+      "volleymanager.volleyball.ch è l'unica fonte autorevole",
+    safeModeWarningPoint2:
+      "Verifica sempre le tue modifiche sul sito ufficiale VolleyManager",
+    safeModeWarningPoint3:
+      "VolleyKit non si assume alcuna responsabilità per eventuali errori",
+    safeModeConfirmButton: "Ho capito, disattiva",
+    safeModeDangerous: "Le operazioni pericolose sono abilitate",
+    safeModeBlocked:
+      "Questa operazione è bloccata in modalità sicura. Disattiva la modalità sicura nelle Impostazioni per procedere.",
+    privacy: "Privacy",
+    privacyNoCollection:
+      "VolleyKit non raccoglie né memorizza alcun dato personale.",
+    privacyDirectComm:
+      "Tutti i dati fluiscono direttamente tra il tuo browser e i server di Swiss Volley.",
+    privacyNoAnalytics: "Nessun tracciamento, analisi o telemetria.",
+    about: "Informazioni",
+    version: "Versione",
+    platform: "Piattaforma",
+    openWebsite: "Apri sito VolleyManager",
+    roles: "Ruoli",
+    dataSource: "Dati da volleymanager.volleyball.ch",
+    disclaimer:
+      "App non ufficiale per uso personale. Tutti i dati sono proprietà di Swiss Volley.",
+    updates: "Aggiornamenti",
+    checkForUpdates: "Verifica aggiornamenti",
+    checking: "Verifica in corso...",
+    upToDate: "L'app è aggiornata",
+    updateAvailable: "Aggiornamento disponibile",
+    lastChecked: "Ultimo controllo",
+    updateNow: "Aggiorna ora",
+    updateCheckFailed: "Verifica aggiornamenti fallita",
+  },
+  validation: {
+    homeRoster: "Rosa di casa",
+    awayRoster: "Rosa ospite",
+    scorer: "Segnapunti",
+    scoresheet: "Referto",
+    homeRosterPlaceholder:
+      "La verifica della rosa di casa sarà disponibile qui.",
+    awayRosterPlaceholder:
+      "La verifica della rosa ospite sarà disponibile qui.",
+    scorerPlaceholder: "L'identificazione del segnapunti sarà disponibile qui.",
+    scoresheetPlaceholder: "Il caricamento del referto sarà disponibile qui.",
+    addPlayer: "Aggiungi giocatore",
+    searchPlayers: "Cerca giocatori...",
+    noPlayersFound: "Nessun giocatore trovato",
+    loadPlayersError: "Impossibile caricare i giocatori",
+    playerAlreadyAdded: "Già nella rosa",
+    jerseyNumber: "Maglia #",
+    license: "Licenza",
+    roster: {
+      addPlayer: "Aggiungi giocatore",
+      removePlayer: "Rimuovi giocatore",
+      undoRemoval: "Annulla rimozione",
+      newlyAdded: "Nuovo",
+      captain: "Capitano",
+      libero: "Libero",
+      emptyRoster: "Nessun giocatore nella rosa",
+      loadingRoster: "Caricamento rosa...",
+      errorLoading: "Caricamento rosa fallito",
+      playerCount: "{count} giocatori",
+    },
+    scorerSearch: {
+      searchPlaceholder: "Cerca segnapunti per nome...",
+      searchHint:
+        "Inserisci il nome (es. 'Müller' o 'Hans Müller') o aggiungi l'anno di nascita (es. 'Müller 1985')",
+      searchError: "Ricerca segnapunti fallita",
+      noScorerSelected:
+        "Nessun segnapunti selezionato. Usa la ricerca sopra per trovare e selezionare un segnapunti.",
+      noScorersFound: "Nessun segnapunti trovato",
+      searchResults: "Risultati della ricerca",
+      resultsCount: "{count} risultati trovati",
+      resultsCountOne: "1 risultato trovato",
+    },
+    scoresheetUpload: {
+      title: "Carica referto",
+      description: "Carica una foto o una scansione del referto fisico",
+      acceptedFormats: "JPEG, PNG o PDF",
+      maxFileSize: "Max 10 MB",
+      selectFile: "Seleziona file",
+      takePhoto: "Scatta foto",
+      uploading: "Caricamento...",
+      uploadComplete: "Caricamento completato",
+      replace: "Sostituisci",
+      remove: "Rimuovi",
+      fileTooLarge: "Il file è troppo grande. Dimensione massima: 10 MB.",
+      invalidFileType: "Tipo di file non valido. Usa JPEG, PNG o PDF.",
+      demoModeNote: "Modalità demo: i caricamenti sono simulati",
+      previewAlt: "Anteprima referto",
+    },
+    state: {
+      unsavedChangesTitle: "Modifiche non salvate",
+      unsavedChangesMessage:
+        "Hai modifiche non salvate. Sei sicuro di volerle scartare?",
+      continueEditing: "Continua a modificare",
+      discardChanges: "Scarta le modifiche",
+      saveSuccess: "Validazione salvata con successo",
+      saveError: "Impossibile salvare la validazione",
+      saveDisabledTooltip:
+        "Completa tutti i pannelli richiesti (Rosa di casa, Rosa ospite, Segnapunti) per salvare",
+    },
+  },
+};
+
+export default it;
