+++ web-app/src/i18n/types.ts
@@ -1,0 +1,234 @@
+/**
+ * Translation keys organized by feature.
+ * Shared type definition for all locale files.
+ */
+export interface Translations {
+  common: {
+    loading: string;
+    error: string;
+    retry: string;
+    cancel: string;
+    save: string;
+    close: string;
+    confirm: string;
+    noResults: string;
+    today: string;
+    tomorrow: string;
+    home: string;
+    away: string;
+    men: string;
+    women: string;
+    match: string;
+    dateTime: string;
+    location: string;
+    position: string;
+    requiredLevel: string;
+    demoModeBanner: string;
+    optional: string;
+  };
+  auth: {
+    login: string;
+    logout: string;
+    username: string;
+    password: string;
+    loginButton: string;
+    loggingIn: string;
+    invalidCredentials: string;
+    sessionExpired: string;
+    subtitle: string;
+    or: string;
+    demoMode: string;
+    loginInfo: string;
+    privacyNote: string;
+    loadingDemo: string;
+  };
+  occupations: {
+    referee: string;
+    player: string;
+    clubAdmin: string;
+    associationAdmin: string;
+  };
+  assignments: {
+    title: string;
+    upcoming: string;
+    past: string;
+    validationClosed: string;
+    loading: string;
+    noAssignments: string;
+    noUpcomingTitle: string;
+    noUpcomingDescription: string;
+    noClosedTitle: string;
+    noClosedDescription: string;
+    confirmed: string;
+    pending: string;
+    cancelled: string;
+    editCompensation: string;
+    validateGame: string;
+    kilometers: string;
+    reason: string;
+    reasonPlaceholder: string;
+    homeScore: string;
+    awayScore: string;
+    numberOfSets: string;
+    gameReportNotAvailable: string;
+  };
+  compensations: {
+    title: string;
+    noCompensations: string;
+    paid: string;
+    unpaid: string;
+    pending: string;
+    total: string;
+    all: string;
+    received: string;
+    loading: string;
+    noCompensationsTitle: string;
+    noCompensationsDescription: string;
+    noPaidTitle: string;
+    noPaidDescription: string;
+    noUnpaidTitle: string;
+    noUnpaidDescription: string;
+    errorLoading: string;
+    pdfNotAvailableDemo: string;
+    pdfDownloadFailed: string;
+  };
+  exchange: {
+    title: string;
+    noExchanges: string;
+    apply: string;
+    withdraw: string;
+    open: string;
+    applied: string;
+    closed: string;
+    all: string;
+    myApplications: string;
+    takeOverTitle: string;
+    takeOverConfirm: string;
+    takeOverButton: string;
+    removeTitle: string;
+    removeConfirm: string;
+    removeButton: string;
+    filterByLevel: string;
+    noExchangesAtLevel: string;
+    noOpenExchangesTitle: string;
+    noOpenExchangesDescription: string;
+    noApplicationsTitle: string;
+    noApplicationsDescription: string;
+  };
+  positions: {
+    "head-one": string;
+    "head-two": string;
+    "linesman-one": string;
+    "linesman-two": string;
+    "linesman-three": string;
+    "linesman-four": string;
+    "standby-head": string;
+    "standby-linesman": string;
+  };
+  nav: {
+    assignments: string;
+    compensations: string;
+    exchange: string;
+    settings: string;
+  };
+  settings: {
+    title: string;
+    profile: string;
+    language: string;
+    safeMode: string;
+    safeModeDescription: string;
+    safeModeEnabled: string;
+    safeModeDisabled: string;
+    safeModeWarningTitle: string;
+    safeModeWarningMessage: string;
+    safeModeWarningPoint1: string;
+    safeModeWarningPoint2: string;
+    safeModeWarningPoint3: string;
+    safeModeConfirmButton: string;
+    safeModeDangerous: string;
+    safeModeBlocked: string;
+    privacy: string;
+    privacyNoCollection: string;
+    privacyDirectComm: string;
+    privacyNoAnalytics: string;
+    about: string;
+    version: string;
+    platform: string;
+    openWebsite: string;
+    roles: string;
+    dataSource: string;
+    disclaimer: string;
+    updates: string;
+    checkForUpdates: string;
+    checking: string;
+    upToDate: string;
+    updateAvailable: string;
+    lastChecked: string;
+    updateNow: string;
+    updateCheckFailed: string;
+  };
+  validation: {
+    homeRoster: string;
+    awayRoster: string;
+    scorer: string;
+    scoresheet: string;
+    homeRosterPlaceholder: string;
+    awayRosterPlaceholder: string;
+    scorerPlaceholder: string;
+    scoresheetPlaceholder: string;
+    addPlayer: string;
+    searchPlayers: string;
+    noPlayersFound: string;
+    loadPlayersError: string;
+    playerAlreadyAdded: string;
+    jerseyNumber: string;
+    license: string;
+    roster: {
+      addPlayer: string;
+      removePlayer: string;
+      undoRemoval: string;
+      newlyAdded: string;
+      captain: string;
+      libero: string;
+      emptyRoster: string;
+      loadingRoster: string;
+      errorLoading: string;
+      playerCount: string;
+    };
+    scorerSearch: {
+      searchPlaceholder: string;
+      searchHint: string;
+      searchError: string;
+      noScorerSelected: string;
+      noScorersFound: string;
+      searchResults: string;
+      resultsCount: string;
+      resultsCountOne: string;
+    };
+    scoresheetUpload: {
+      title: string;
+      description: string;
+      acceptedFormats: string;
+      maxFileSize: string;
+      selectFile: string;
+      takePhoto: string;
+      uploading: string;
+      uploadComplete: string;
+      replace: string;
+      remove: string;
+      fileTooLarge: string;
+      invalidFileType: string;
+      demoModeNote: string;
+      previewAlt: string;
+    };
+    state: {
+      unsavedChangesTitle: string;
+      unsavedChangesMessage: string;
+      continueEditing: string;
+      discardChanges: string;
+      saveSuccess: string;
+      saveError: string;
+      saveDisabledTooltip: string;
+    };
+  };
+}
